+++ src/main/java/com/haruhiism/bbs/controller/CommentController.java
@@ -1,6 +1,8 @@
 package com.haruhiism.bbs.controller;
 
 import com.haruhiism.bbs.command.comment.CommentRemoveRequestCommand;
+import com.haruhiism.bbs.command.comment.CommentRemoveRequestValidationGroup;
+import com.haruhiism.bbs.command.comment.CommentRemoveSubmitValidationGroup;
 import com.haruhiism.bbs.command.comment.CommentSubmitCommand;
 import com.haruhiism.bbs.domain.authentication.LoginSessionInfo;
 import com.haruhiism.bbs.domain.dto.AuthDTO;
@@ -9,16 +11,17 @@ import com.haruhiism.bbs.exception.auth.AuthenticationFailedException;
 import com.haruhiism.bbs.service.comment.CommentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
-import javax.validation.constraints.Positive;
 
 @Controller
 @RequestMapping("/comment")
@@ -30,7 +33,13 @@ public class CommentController {
 
     @PostMapping("/create")
     public String createComment(@Valid CommentSubmitCommand command,
+                                BindingResult bindingResult,
                                 HttpServletRequest request) {
+
+        if (bindingResult.hasErrors()) {
+            return "redirect:/board/read?id=" + command.getArticleID();
+        }
+
         LoginSessionInfo loginSessionInfo = getLoginSessionInfoFromHttpSession(request.getSession(false));
 
         commentService.createComment(BoardCommentDTO.builder()
@@ -45,33 +54,47 @@ public class CommentController {
 
 
     @GetMapping("/remove")
-    public String requestRemoveComment(Model model,
-                                       @RequestParam("id") @Positive Long commentID,
+    public String requestRemoveComment(@ModelAttribute("command") @Validated(CommentRemoveRequestValidationGroup.class) CommentRemoveRequestCommand command,
+                                       BindingResult bindingResult,
                                        HttpServletRequest request){
 
-        BoardCommentDTO commentDTO = commentService.readComment(commentID);
+        if (bindingResult.hasErrors()) {
+            return "redirect:/board/read?id=" + command.getId();
+        }
+
+        BoardCommentDTO commentDTO = commentService.readComment(command.getId());
         if (!commentDTO.getUserId().isBlank()) {
             LoginSessionInfo loginSessionInfo = getLoginSessionInfoFromHttpSession(request.getSession(false));
-            if(loginSessionInfo == null) throw new AuthenticationFailedException();
+            if(loginSessionInfo != null) {
+                commentService.deleteComment(command.getId(), AuthDTO.builder().loginSessionInfo(loginSessionInfo).build());
+            }
 
-            commentService.deleteComment(commentID, AuthDTO.builder().loginSessionInfo(loginSessionInfo).build());
-            return String.format("redirect:/board/read?id=%d",commentDTO.getArticleId());
+            return "redirect:/board/read?id=" + commentDTO.getArticleId();
         } else {
-            model.addAttribute("commentID", commentID);
             return "comment/removeRequest";
         }
     }
 
     @PostMapping("/remove")
-    public String submitRemoveComment(@Valid CommentRemoveRequestCommand command) {
-        BoardCommentDTO commentDTO = commentService.readComment(command.getId());
-        commentService.deleteComment(command.getId(), AuthDTO.builder().rawPassword(command.getPassword()).build());
-        return String.format("redirect:/board/read?id=%d", commentDTO.getArticleId());
+    public String submitRemoveComment(@ModelAttribute("command") @Validated(CommentRemoveSubmitValidationGroup.class) CommentRemoveRequestCommand command,
+                                      BindingResult bindingResult) {
+
+        if(bindingResult.hasErrors()){
+            return "comment/removeRequest";
+        }
+
+        try {
+            BoardCommentDTO commentDTO = commentService.readComment(command.getId());
+            commentService.deleteComment(command.getId(), AuthDTO.builder().rawPassword(command.getPassword()).build());
+            return "redirect:/board/read?id=" + commentDTO.getArticleId();
+        } catch (AuthenticationFailedException exception) {
+            bindingResult.addError(new FieldError("command", "password", "Password not matched."));
+            return "comment/removeRequest";
+        }
     }
 
     private LoginSessionInfo getLoginSessionInfoFromHttpSession(HttpSession session) {
         if(session == null) return null;
-        String sessionAuthAttribute = "loginSessionInfo";
-        return (LoginSessionInfo) session.getAttribute(sessionAuthAttribute);
+        return (LoginSessionInfo) session.getAttribute("loginSessionInfo");
     }
 }
