+++ src/main/java/com/haruhiism/bbs/controller/ManageController.java
@@ -1,15 +1,15 @@
 package com.haruhiism.bbs.controller;
 
-import com.haruhiism.bbs.command.account.AccountListCommand;
+import com.haruhiism.bbs.command.DateBasedListCommand;
 import com.haruhiism.bbs.command.article.ArticleListCommand;
-import com.haruhiism.bbs.command.comment.CommentListCommand;
+import com.haruhiism.bbs.command.manage.AccountListCommand;
 import com.haruhiism.bbs.command.manage.AccountManagementCommand;
 import com.haruhiism.bbs.command.manage.BoardManagementCommand;
+import com.haruhiism.bbs.command.manage.CommentListCommand;
+import com.haruhiism.bbs.domain.AccountLevel;
 import com.haruhiism.bbs.domain.authentication.LoginSessionInfo;
-import com.haruhiism.bbs.domain.dto.BoardAccountDTO;
-import com.haruhiism.bbs.domain.dto.BoardAccountsDTO;
-import com.haruhiism.bbs.domain.dto.BoardArticlesDTO;
-import com.haruhiism.bbs.domain.dto.BoardCommentsDTO;
+import com.haruhiism.bbs.domain.dto.*;
+import com.haruhiism.bbs.service.account.AccountService;
 import com.haruhiism.bbs.service.manage.AccountManagerService;
 import com.haruhiism.bbs.service.manage.ArticleManagerService;
 import com.haruhiism.bbs.service.manage.CommentManagerService;
@@ -17,6 +17,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -25,7 +26,9 @@ import org.springframework.web.bind.annotation.SessionAttribute;
 import javax.validation.Valid;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
+import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 
 @Controller
 @RequestMapping("/manage")
@@ -37,7 +40,43 @@ public class ManageController {
     private final CommentManagerService commentManagerService;
     private final AccountManagerService accountManagerService;
 
+    private final AccountService accountService;
+
     private final String sessionAuthAttribute = "loginSessionInfo";
+
+
+    private Map<String, String> generateConsoleLinksByAccountLevel(List<AccountLevel> levels){
+        Map<String, String> links = new LinkedHashMap<>();
+        links.put("BOARD", "/board/list");
+        links.put("CONSOLE", "/manage/console");
+        for (AccountLevel level : levels) {
+            switch(level){
+                case ACCOUNT_MANAGER:
+                    links.put("ACCOUNTS", "/manage/console/account");
+                    break;
+
+                case BOARD_MANAGER:
+                    links.put("ARTICLES", "/manage/console/article");
+                    links.put("COMMENTS", "/manage/console/comment");
+                    break;
+            }
+        }
+
+        return links;
+    }
+
+    private void addPaginationToModel(Model model, DTOContainer result){
+        model.addAttribute("currentPage", result.getCurrentPage());
+        model.addAttribute("totalPages", result.getTotalPages());
+    }
+
+    private void addDateCommandToModel(Model model, DateBasedListCommand command){
+        model.addAttribute("pageSize", command.getPageSize());
+        model.addAttribute("keyword", command.getKeyword());
+        model.addAttribute("from", command.getFrom());
+        model.addAttribute("to", command.getTo());
+        model.addAttribute("betweenDates", command.isBetweenDates());
+    }
 
     @GetMapping("/console")
     public String showManagementPage(Model model,
@@ -50,18 +89,27 @@ public class ManageController {
 
         model.addAttribute("signedAccounts", accountManagerService.countAllAccounts());
 
-        model.addAttribute("loginSessionInfo", loginSessionInfo);
-        model.addAttribute("levels", accountManagerService.getLevelOfAccount(
-                BoardAccountDTO.builder().userId(loginSessionInfo.getUserID()).build()));
+        model.addAttribute("userInfo", loginSessionInfo);
+        List<AccountLevel> levels = accountService.getAccountLevels(BoardAccountDTO.builder().userId(loginSessionInfo.getUserID()).build()).getLevels();
+        model.addAttribute("links", generateConsoleLinksByAccountLevel(levels));
 
         return "admin/management-console";
     }
 
 
     @GetMapping("/console/article")
-    public String articleManagementPage(@Valid ArticleListCommand command, Model model){
+    public String articleManagementPage(@Valid ArticleListCommand command,
+                                        BindingResult bindingResult,
+                                        Model model,
+                                        @SessionAttribute(sessionAuthAttribute)LoginSessionInfo loginSessionInfo){
+
+        if(bindingResult.hasErrors()){
+            return "redirect:/manage/console/article";
+        }
+
         BoardArticlesDTO result;
         if(command.getKeyword().isBlank()) {
+            // TODO: ArticleService에 파라미터로 검색 조건을 설정하는 방식으로 통합?
             result = articleManagerService.readArticles(
                     command.getPageNum(), command.getPageSize(),
                     command.isBetweenDates() ? LocalDateTime.of(command.getFrom(), LocalTime.of(0, 0)) : LocalDateTime.MIN,
@@ -73,34 +121,34 @@ public class ManageController {
                     command.isBetweenDates() ? LocalDateTime.of(command.getTo(), LocalTime.of(0, 0)) : LocalDateTime.now());
         }
 
-        model.addAttribute("currentPage", result.getCurrentPage());
-        model.addAttribute("pageSize", command.getPageSize());
-        model.addAttribute("totalPage", result.getTotalPages());
+        addPaginationToModel(model, result);
+        addDateCommandToModel(model, command);
 
         model.addAttribute("articles", result.getBoardArticles());
         model.addAttribute("commentSizes", result.getBoardArticleCommentSize());
-
-        model.addAttribute("keyword", command.getKeyword());
         model.addAttribute("mode", command.getMode().name());
 
-        model.addAttribute("from", command.getFrom());
-        model.addAttribute("to", command.getTo());
-        model.addAttribute("betweenDates", command.isBetweenDates());
+        List<AccountLevel> levels = accountService.getAccountLevels(BoardAccountDTO.builder().userId(loginSessionInfo.getUserID()).build()).getLevels();
+        model.addAttribute("links", generateConsoleLinksByAccountLevel(levels));
 
         return "admin/article-console";
     }
 
     @PostMapping("/console/article")
-    public String submitArticleManagements(@Valid BoardManagementCommand command) {
-        List<Long> articleIds = command.getTarget();
-        switch (command.getOperation()) {
-            case DELETE:
-                articleManagerService.deleteArticles(articleIds);
-                break;
+    public String submitArticleManagements(@Valid BoardManagementCommand command,
+                                           BindingResult bindingResult) {
 
-            case RESTORE:
-                articleManagerService.restoreArticles(articleIds);
-                break;
+        if (!bindingResult.hasErrors()) {
+            List<Long> articleIds = command.getTarget();
+            switch (command.getOperation()) {
+                case DELETE:
+                    articleManagerService.deleteArticles(articleIds);
+                    break;
+
+                case RESTORE:
+                    articleManagerService.restoreArticles(articleIds);
+                    break;
+            }
         }
 
         return "redirect:/manage/console/article";
@@ -109,7 +157,14 @@ public class ManageController {
 
     @GetMapping("/console/comment")
     public String commentManagementPage(@Valid CommentListCommand command,
-                                        Model model){
+                                        BindingResult bindingResult,
+                                        Model model,
+                                        @SessionAttribute(sessionAuthAttribute)LoginSessionInfo loginSessionInfo){
+
+        if (bindingResult.hasErrors()) {
+            return "redirect:/manage/console/comment";
+        }
+
         BoardCommentsDTO result;
         if(command.getKeyword().isBlank()){
             result = commentManagerService.readComments(
@@ -123,42 +178,47 @@ public class ManageController {
                     command.isBetweenDates() ? LocalDateTime.of(command.getTo(), LocalTime.of(0, 0)) : LocalDateTime.now());
         }
 
-        model.addAttribute("currentPage", result.getCurrentPage());
-        model.addAttribute("totalPage", result.getTotalPages());
-        model.addAttribute("pageSize", command.getPageSize());
-
+        addPaginationToModel(model, result);
+        addDateCommandToModel(model, command);
         model.addAttribute("comments", result.getBoardComments());
-
-        model.addAttribute("keyword", command.getKeyword());
         model.addAttribute("mode", command.getMode().name());
 
-        model.addAttribute("from", command.getFrom());
-        model.addAttribute("to", command.getTo());
-        model.addAttribute("betweenDates", command.isBetweenDates());
+        List<AccountLevel> levels = accountService.getAccountLevels(BoardAccountDTO.builder().userId(loginSessionInfo.getUserID()).build()).getLevels();
+        model.addAttribute("links", generateConsoleLinksByAccountLevel(levels));
 
         return "admin/comment-console";
     }
 
     @PostMapping("/console/comment")
-    public String submitCommentManagements(@Valid BoardManagementCommand command){
-        List<Long> commentIds = command.getTarget();
-        switch (command.getOperation()) {
-            case DELETE:
-                commentManagerService.deleteComments(commentIds);
-                break;
+    public String submitCommentManagements(@Valid BoardManagementCommand command,
+                                           BindingResult bindingResult){
 
-            case RESTORE:
-                commentManagerService.restoreComments(commentIds);
-                break;
+        if (!bindingResult.hasErrors()) {
+            List<Long> commentIds = command.getTarget();
+            switch (command.getOperation()) {
+                case DELETE:
+                    commentManagerService.deleteComments(commentIds);
+                    break;
+
+                case RESTORE:
+                    commentManagerService.restoreComments(commentIds);
+                    break;
+            }
         }
 
         return "redirect:/manage/console/comment";
-
     }
 
 
     @GetMapping("/console/account")
-    public String accountManagementPage(@Valid AccountListCommand command, Model model){
+    public String accountManagementPage(@Valid AccountListCommand command,
+                                        BindingResult bindingResult,
+                                        Model model,
+                                        @SessionAttribute(sessionAuthAttribute)LoginSessionInfo loginSessionInfo){
+
+        if (bindingResult.hasErrors()) {
+            return "redirect:/manage/console/account";
+        }
 
         BoardAccountsDTO accounts;
         if(command.getKeyword().isBlank()){
@@ -173,25 +233,27 @@ public class ManageController {
                     command.isBetweenDates() ? LocalDateTime.of(command.getTo(), LocalTime.of(0,0)) : LocalDateTime.now());
         }
 
+        addPaginationToModel(model, accounts);
+        addDateCommandToModel(model, command);
+
         model.addAttribute("accounts", accounts.getAccounts());
-
-        model.addAttribute("currentPage", accounts.getCurrentPage());
-        model.addAttribute("totalPage", accounts.getTotalPage());
-        model.addAttribute("pageSize", command.getPageSize());
-
-        model.addAttribute("keyword", command.getKeyword());
         model.addAttribute("mode", command.getMode().name());
 
-        model.addAttribute("from", command.getFrom());
-        model.addAttribute("to", command.getTo());
-        model.addAttribute("betweenDates", command.isBetweenDates());
+        List<AccountLevel> levels = accountService.getAccountLevels(BoardAccountDTO.builder().userId(loginSessionInfo.getUserID()).build()).getLevels();
+        model.addAttribute("links", generateConsoleLinksByAccountLevel(levels));
 
         return "admin/account-console";
     }
 
 
     @PostMapping("/console/account")
-    public String submitAccountManagements(@Valid AccountManagementCommand command) {
+    public String submitAccountManagements(@Valid AccountManagementCommand command,
+                                           BindingResult bindingResult) {
+
+        if (bindingResult.hasErrors()) {
+            return "redirect:/manage/console/account";
+        }
+
         List<Long> target = command.getTarget();
         switch (command.getOperation()) {
             case INVALIDATE:
